Setup Netlify
This sets up the Netlify configuration for this site, and allows the
beginning of the migration from both GitHub pages to the standard way we
deploy website properties across Apollo: Netlify.

The site is already in Netlify at:
  https://app.netlify.com/sites/apollo-specs-core/

While the website-router is currently used for the specs.apollo.dev domain,
that was only a temporary configuration.  This root specs site is intended to
become the "proxier" for the other specs, hence the proxy rules that lie
within this commit.